Refactor tests to minimize impact on existing test classes.
